+++ extensions/cloudflare/src/service.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosInstance } from 'axios';
-import {Cache} from "@raycast/api";
+import { Cache } from '@raycast/api';
 
 interface Response<T> {
   result: T;
@@ -183,7 +183,9 @@ class Service {
     if (this.cache.has('accounts')) {
       data = JSON.parse(this.cache.get('accounts')!) as Response<AccountItem[]>;
     } else {
-      const response = await this.client.get<Response<AccountItem[]>>('accounts');
+      const response = await this.client.get<Response<AccountItem[]>>(
+        'accounts',
+      );
       data = response.data;
       this.cache.set('accounts', JSON.stringify(data));
     }